requirements.txt
Django
djangorestframework
mongoengine
python-telegram-bot
pandas
numpy
